graphics working

+++ cpu_helpers.h
@@ -22,7 +22,7 @@ void getBits(uint8_t *, uint8_t);
 
 void writeBits(uint8_t *, uint8_t , uint8_t , chip8 *);
 
-void loadSprites(uint16_t *, uint16_t *);
+void loadSprites(uint8_t *, uint8_t *);
 
 void countTimer(chip8 *);		//for counting down the delay timer
